requirements.txt
django==3.1.7
requests==2.25.1
django-extensions==3.1.1
django-jsonfield==1.4.1
